+++ examples/autoscale_fgt_as_hub/versions.tf
@@ -1,5 +1,5 @@
 terraform {
-  required_version = ">=0.13, < 2.0.0"
+  required_version = ">=1.9.0, < 2.0.0"
   required_providers {
     google = {
       version = ">= 5.0, <8.0"